feat: reduce memory usage
