+++ vulnz/src/main/java/io/github/jeremylong/vulnz/cli/services/NvdMirrorService.java
@@ -1,0 +1,362 @@
+/*
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright (c) 2025 Jeremy Long. All Rights Reserved.
+ */
+package io.github.jeremylong.vulnz.cli.services;
+
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectWriter;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import io.github.jeremylong.openvulnerability.client.nvd.CveApiJson20;
+import io.github.jeremylong.openvulnerability.client.nvd.DefCveItem;
+import io.github.jeremylong.openvulnerability.client.nvd.NvdCveClient;
+import io.github.jeremylong.openvulnerability.client.nvd.NvdCveClientBuilder;
+import io.github.jeremylong.vulnz.cli.cache.CacheException;
+import io.github.jeremylong.vulnz.cli.cache.CacheProperties;
+import io.github.jeremylong.vulnz.cli.cache.CacheUpdateException;
+import io.github.jeremylong.vulnz.cli.ui.IProgressMonitor;
+import io.github.jeremylong.vulnz.cli.ui.JlineShutdownHook;
+import io.github.jeremylong.vulnz.cli.ui.ProgressMonitor;
+import io.github.jeremylong.vulnz.cli.util.HexUtil;
+import io.prometheus.metrics.core.metrics.Gauge;
+import it.unimi.dsi.fastutil.objects.Object2ObjectMap;
+import it.unimi.dsi.fastutil.objects.Object2ObjectOpenHashMap;
+import org.apache.commons.io.FileUtils;
+import org.apache.commons.io.output.CountingOutputStream;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.BufferedOutputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
+import java.io.IOException;
+import java.io.OutputStreamWriter;
+import java.io.PrintWriter;
+import java.nio.charset.StandardCharsets;
+import java.security.DigestOutputStream;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+import java.time.Year;
+import java.time.ZoneOffset;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
+import java.time.temporal.ChronoUnit;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.List;
+import java.util.Optional;
+import java.util.zip.GZIPInputStream;
+import java.util.zip.GZIPOutputStream;
+
+public class NvdMirrorService {
+    /**
+     * Reference to the logger.
+     */
+    private static final Logger LOG = LoggerFactory.getLogger(NvdMirrorService.class);
+    private static final int BUFFER_SIZE = 8192;
+    private static final Gauge CVE_LOAD_COUNTER = Gauge.builder().name("cve_load_counter")
+            .help("Total number of loaded cve's").register();
+    private static final Gauge CVE_COUNTER = Gauge.builder().name("cve_counter").help("Total number of cached cve's")
+            .register();
+
+    private boolean interactive;
+    private CacheProperties properties;
+    private ObjectMapper objectMapper;
+    private Object2ObjectOpenHashMap<String, Object2ObjectOpenHashMap<String, DefCveItem>> cves;
+    private NvdCveClientBuilder nvdCveClientBuilder;
+
+    public NvdMirrorService(File propertyFile, String prefix, NvdCveClientBuilder nvdCveClientBuilder,
+            boolean interactive) {
+        this.properties = new CacheProperties(propertyFile);
+        if (prefix != null) {
+            properties.set("prefix", prefix);
+        }
+        this.nvdCveClientBuilder = nvdCveClientBuilder;
+        this.interactive = interactive;
+        objectMapper = new ObjectMapper();
+        objectMapper.registerModule(new JavaTimeModule());
+        cves = new Object2ObjectOpenHashMap<>();
+        populateKeys(cves);
+    }
+
+    public int process(String apiKey) {
+        NvdCveClientBuilder builder = nvdCveClientBuilder.clone();
+        checkIfHttpCacheExists();
+        if (properties.has("lastModifiedDate")) {
+            ZonedDateTime start = properties.getTimestamp("lastModifiedDate");
+            ZonedDateTime end = start.plusDays(120);
+            if (end.compareTo(ZonedDateTime.now()) > 0) {
+                builder.withLastModifiedFilter(start, end);
+            } else {
+                LOG.warn("Requesting the entire set of NVD CVE data via the api "
+                        + "as the cache was last updated over 120 days ago");
+            }
+        }
+
+        try {
+            int exitCode = processMirrorRequest(builder);
+            if (exitCode == 0 /* success */) {
+                boolean usedExistingCache = Boolean
+                        .parseBoolean(properties.get(CacheProperties.USING_EXISTING_CACHE, "false"));
+                performCleanup(properties);
+                properties.save();
+                if (usedExistingCache) {
+                    exitCode = updateMirrorWithLast24Hours(apiKey);
+                }
+            }
+            return exitCode;
+        } catch (CacheException ex) {
+            LOG.error(ex.getMessage(), ex);
+        }
+        return 1;
+    }
+
+    private Integer processMirrorRequest(NvdCveClientBuilder builder) {
+
+        determineHttpCachingStrategy(builder);
+
+        ZonedDateTime lastModified = null;
+        try {
+            lastModified = downloadAllUpdates(builder, cves);
+        } catch (CacheUpdateException e) {
+            LOG.error("NVD Cache Update Failed", e);
+            return 100;
+        }
+        if (lastModified != null) {
+            properties.set("lastModifiedDate", lastModified);
+        }
+        // update cache
+        writeJsonMirror(cves, lastModified);
+        return 0;
+    }
+
+    private int updateMirrorWithLast24Hours(String apiKey) {
+        NvdCveClientBuilder builder;
+        int exitCode;
+        LOG.info(
+                "Used previously cached HTTP Responses from the NVD; running the update again to obtain entries modified in the last 24 hours");
+        builder = nvdCveClientBuilder.clone();
+        ZonedDateTime start = ZonedDateTime.now(ZoneOffset.UTC).minusHours(24);
+        ZonedDateTime end = start.plusDays(120);
+        builder.withLastModifiedFilter(start, end);
+        exitCode = processMirrorRequest(builder);
+        if (exitCode == 0) {
+            properties.save();
+        }
+        return exitCode;
+    }
+
+    private void determineHttpCachingStrategy(NvdCveClientBuilder builder) {
+        if (Boolean.parseBoolean(properties.get(CacheProperties.IS_NEW, "false"))) {
+            final File httpCacheDir = getHttpCacheDir(properties);
+            if (LOG.isDebugEnabled()) {
+                LOG.debug("Using http cache directory {}", httpCacheDir);
+            }
+            builder.withCacheDirectory(httpCacheDir.toPath());
+        }
+    }
+
+    private void checkIfHttpCacheExists() {
+        if (Boolean.parseBoolean(properties.get(CacheProperties.IS_NEW, "false"))) {
+            File cacheDir = properties.getDirectory();
+            if (cacheDir != null && cacheDir.exists() && cacheDir.isDirectory()) {
+                File[] files = cacheDir.listFiles();
+                if (files != null && files.length > 0 && Arrays.stream(files)
+                        .anyMatch(file -> file.lastModified() >= System.currentTimeMillis() - 72000000)) {
+                    properties.set(CacheProperties.USING_EXISTING_CACHE, "true");
+                }
+            }
+        }
+    }
+
+    private static File getHttpCacheDir(CacheProperties properties) {
+        return new File(properties.getDirectory(), ".nvd_requests");
+    }
+
+    private static void populateKeys(
+            Object2ObjectOpenHashMap<String, Object2ObjectOpenHashMap<String, DefCveItem>> cves) {
+        cves.put("modified", new Object2ObjectOpenHashMap<>());
+        for (int i = 2002; i <= Year.now().getValue(); i++) {
+            cves.put(Integer.toString(i), new Object2ObjectOpenHashMap<>());
+        }
+    }
+
+    private ZonedDateTime downloadAllUpdates(NvdCveClientBuilder builder,
+            Object2ObjectOpenHashMap<String, Object2ObjectOpenHashMap<String, DefCveItem>> cves)
+            throws CacheUpdateException {
+        ZonedDateTime lastModified = null;
+        int count = 0;
+        // retrieve from NVD API
+        try (NvdCveClient api = builder.build(); IProgressMonitor monitor = new ProgressMonitor(interactive, "NVD")) {
+            Runtime.getRuntime().addShutdownHook(new JlineShutdownHook());
+            while (api.hasNext()) {
+                Collection<DefCveItem> data = api.next();
+                collectCves(cves, data);
+                lastModified = api.getLastUpdated();
+                count += data.size();
+                CVE_LOAD_COUNTER.set(count);
+                writeJsonMirror(cves, null, 2000);
+                monitor.updateProgress("NVD", count, api.getTotalAvailable());
+            }
+        } catch (Exception ex) {
+            LOG.debug("\nERROR", ex);
+            throw new CacheUpdateException("Unable to complete NVD cache update due to error: " + ex.getMessage());
+        }
+        CVE_COUNTER.set(cves.values().stream().mapToLong(Object2ObjectOpenHashMap::size).sum());
+        return lastModified;
+    }
+
+    private String getNvdYear(DefCveItem item) {
+        int year = item.getCve().getPublished().getYear();
+        if (year < 2002) {
+            year = 2002;
+        }
+        return Integer.toString(year);
+    }
+
+    private void collectCves(Object2ObjectOpenHashMap<String, Object2ObjectOpenHashMap<String, DefCveItem>> cves,
+            Collection<DefCveItem> vulnerabilities) {
+        synchronized (cves) {
+            for (DefCveItem item : vulnerabilities) {
+                cves.get(getNvdYear(item)).put(item.getCve().getId(), item);
+                if (ChronoUnit.DAYS.between(item.getCve().getLastModified(), ZonedDateTime.now()) <= 7) {
+                    cves.get("modified").put(item.getCve().getId(), item);
+                }
+            }
+        }
+    }
+
+    private void writeJsonMirror(Object2ObjectOpenHashMap<String, Object2ObjectOpenHashMap<String, DefCveItem>> cves,
+            ZonedDateTime lastModified) {
+        writeJsonMirror(cves, lastModified, 0);
+    }
+
+    @SuppressFBWarnings(value = "DM_GC", justification = "Without calling System.gc() the JVM uses 2+GB of memory, calling gc profiling results in less than 1GB")
+    private void writeJsonMirror(Object2ObjectOpenHashMap<String, Object2ObjectOpenHashMap<String, DefCveItem>> cves,
+            ZonedDateTime lastModified, int writeThreshold) {
+        final String format = "NVD_CVE";
+        final String version = "2.0";
+        final String prefix = properties.get("prefix", "nvdcve-");
+
+        List<String> sortedKeys = new ArrayList<>(cves.keySet());
+        sortedKeys.sort((k1, k2) -> k2.compareTo(k1));
+
+        // flush older data after we find an entry that meets the threshold; meaning if 2020
+        // meets the threshold and 2019 still have 500 entries - write the 500 entries
+        boolean flushOlder = false;
+        for (String key : sortedKeys) {
+            Object2ObjectOpenHashMap<String, DefCveItem> cveApiData = cves.get(key);
+            if ((cveApiData.isEmpty() || cveApiData.size() < writeThreshold)
+                    && !(flushOlder && !cveApiData.isEmpty())) {
+                continue;
+            }
+            flushOlder = true;
+            final File file = new File(properties.getDirectory(), prefix + key + ".json.gz");
+            final File meta = new File(properties.getDirectory(), prefix + key + ".meta");
+            final Object2ObjectOpenHashMap<String, DefCveItem> yearData = new Object2ObjectOpenHashMap<>();
+            // Load existing year data if present
+            if (file.isFile()) {
+                try (FileInputStream fis = new FileInputStream(file); GZIPInputStream gzis = new GZIPInputStream(fis)) {
+                    CveApiJson20 data = objectMapper.readValue(gzis, CveApiJson20.class);
+                    boolean isYearData = !"modified".equals(key);
+                    // filter the "modified" data to load only the last 7 days of data
+                    data.getVulnerabilities().stream().filter(cve -> isYearData
+                            || ChronoUnit.DAYS.between(cve.getCve().getLastModified(), ZonedDateTime.now()) <= 7)
+                            .forEach(cve -> yearData.put(cve.getCve().getId(), cve));
+                } catch (IOException exception) {
+                    throw new CacheException("Unable to read cached data: " + file, exception);
+                }
+            }
+            synchronized (cves) {
+                yearData.putAll(cveApiData);
+                cveApiData.clear();
+            }
+
+            List<DefCveItem> vulnerabilities = new ArrayList<DefCveItem>(yearData.values());
+            vulnerabilities.sort((v1, v2) -> {
+                return v1.getCve().getId().compareTo(v2.getCve().getId());
+            });
+            ZonedDateTime timestamp;
+            Optional<ZonedDateTime> maxDate = vulnerabilities.stream().map(v -> v.getCve().getLastModified())
+                    .max(ZonedDateTime::compareTo);
+            if (maxDate.isPresent()) {
+                timestamp = maxDate.get();
+            } else if (lastModified != null) {
+                timestamp = lastModified;
+            } else {
+                timestamp = ZonedDateTime.now();
+            }
+            properties.set("lastModifiedDate." + key, timestamp);
+            CveApiJson20 data = new CveApiJson20(vulnerabilities.size(), 0, vulnerabilities.size(), format, version,
+                    timestamp, vulnerabilities);
+            MessageDigest md;
+            try {
+                md = MessageDigest.getInstance("SHA-256");
+            } catch (NoSuchAlgorithmException e) {
+                throw new CacheException("Unable to calculate sha256 checksum", e);
+            }
+            long byteCount = 0;
+            try (FileOutputStream fileOutputStream = new FileOutputStream(file);
+                    GZIPOutputStream gzipOutputStream = new GZIPOutputStream(fileOutputStream);
+                    DigestOutputStream digestOutputStream = new DigestOutputStream(gzipOutputStream, md);
+                    CountingOutputStream countingOutputStream = new CountingOutputStream(digestOutputStream);
+                    BufferedOutputStream bufferedOutputStream = new BufferedOutputStream(countingOutputStream,
+                            BUFFER_SIZE)) {
+                objectMapper.writeValue(bufferedOutputStream, data);
+                byteCount = countingOutputStream.getByteCount();
+            } catch (IOException ex) {
+                throw new CacheException("Unable to write cached data: " + file, ex);
+            }
+            yearData.clear();
+            vulnerabilities.clear();
+            System.gc();
+            String checksum = HexUtil.getHex(md.digest());
+            try (FileOutputStream fileOutputStream = new FileOutputStream(meta);
+                    OutputStreamWriter osw = new OutputStreamWriter(fileOutputStream, StandardCharsets.UTF_8);
+                    PrintWriter writer = new PrintWriter(osw)) {
+                final String lmd = DateTimeFormatter.ISO_DATE_TIME.format(timestamp);
+                writer.println("lastModifiedDate:" + lmd);
+                writer.println("size:" + byteCount);
+                writer.println("gzSize:" + file.length());
+                writer.println("sha256:" + checksum);
+            } catch (IOException ex) {
+                throw new CacheException("Unable to write cached meta-data: " + file, ex);
+            }
+        }
+    }
+
+    /**
+     * Deletes the HTTP Cache Directory if it exists and is no longer needed, and saves the cache properties.
+     * @param properties the cache properties to save
+     */
+    private void performCleanup(CacheProperties properties) {
+        if (Boolean.parseBoolean(properties.get(CacheProperties.IS_NEW, "false"))) {
+            File cacheDir = getHttpCacheDir(properties);
+            if (cacheDir.exists() && cacheDir.isDirectory()) {
+                try {
+                    FileUtils.deleteDirectory(cacheDir);
+                } catch (IOException ex) {
+                    LOG.error("Unable to delete the HTTP Cache Directory `{}`, error: {}", cacheDir.getAbsoluteFile(),
+                            ex.getMessage());
+                }
+            }
+        }
+        properties.set(CacheProperties.IS_NEW, "false");
+        properties.remove(CacheProperties.USING_EXISTING_CACHE);
+    }
+}
