+++ vulnz/src/main/java/io/github/jeremylong/vulnz/cli/services/NvdService.java
@@ -1,0 +1,118 @@
+/*
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ * Copyright (c) 2025 Jeremy Long. All Rights Reserved.
+ */
+package io.github.jeremylong.vulnz.cli.services;
+
+import com.diogonunes.jcolor.Attribute;
+import com.fasterxml.jackson.core.JsonEncoding;
+import com.fasterxml.jackson.core.JsonFactory;
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializationFeature;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
+import io.github.jeremylong.openvulnerability.client.nvd.DefCveItem;
+import io.github.jeremylong.openvulnerability.client.nvd.NvdCveClient;
+import io.github.jeremylong.openvulnerability.client.nvd.NvdCveClientBuilder;
+import io.github.jeremylong.vulnz.cli.model.BasicOutput;
+import io.github.jeremylong.vulnz.cli.ui.IProgressMonitor;
+import io.github.jeremylong.vulnz.cli.ui.JlineShutdownHook;
+import io.github.jeremylong.vulnz.cli.ui.ProgressMonitor;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+import java.io.IOException;
+import java.util.Collection;
+
+import static com.diogonunes.jcolor.Ansi.colorize;
+
+public class NvdService {
+
+    /**
+     * Reference to the logger.
+     */
+    private static final Logger LOG = LoggerFactory.getLogger(NvdService.class);
+
+    private boolean prettyPrint;
+    private boolean interactive;
+    private NvdCveClientBuilder builder;
+
+    public NvdService(NvdCveClientBuilder builder, boolean prettyPrint, boolean interactive) {
+        this.builder = builder;
+        this.prettyPrint = prettyPrint;
+        this.interactive = interactive;
+    }
+
+    public int process() throws IOException {
+        JsonGenerator jsonOut = getJsonGenerator();
+        int status = 1;
+        jsonOut.writeStartObject();
+        jsonOut.writeFieldName("cves");
+        jsonOut.writeStartArray();
+        BasicOutput output = new BasicOutput();
+        int count = 0;
+        try (NvdCveClient api = builder.build(); IProgressMonitor monitor = new ProgressMonitor(interactive, "NVD")) {
+            Runtime.getRuntime().addShutdownHook(new JlineShutdownHook());
+            while (api.hasNext()) {
+                Collection<DefCveItem> list = api.next();
+                if (list != null) {
+                    count += list.size();
+                }
+                monitor.updateProgress("NVD", count, api.getTotalAvailable());
+                if (list != null) {
+                    output.setSuccess(true);
+                    output.addCount(list.size());
+                    for (DefCveItem c : list) {
+                        jsonOut.writeObject(c.getCve());
+                    }
+                    if (output.getLastModifiedDate() == null
+                            || output.getLastModifiedDate().compareTo(api.getLastUpdated()) < 0) {
+                        output.setLastModifiedDate(api.getLastUpdated());
+                    }
+                } else {
+                    output.setSuccess(false);
+                    output.setReason(String.format("Received HTTP Status Code: %s", api.getLastStatusCode()));
+                }
+            }
+            jsonOut.writeEndArray();
+            jsonOut.writeObjectField("results", output);
+            jsonOut.writeEndObject();
+            jsonOut.close();
+
+            if (!output.isSuccess()) {
+                String msg = String.format("%nFAILED: %s", output.getReason());
+                LOG.info(colorize(msg, Attribute.RED_TEXT()));
+                status = 2;
+            }
+            LOG.info(colorize("\nSUCCESS", Attribute.GREEN_TEXT()));
+            status = 0;
+        } catch (Throwable ex) {
+            LOG.error("\nERROR", ex);
+        }
+        return status;
+    }
+
+    private JsonGenerator getJsonGenerator() throws IOException {
+        ObjectMapper objectMapper = new ObjectMapper();
+        objectMapper.registerModule(new JavaTimeModule());
+        objectMapper.disable(SerializationFeature.WRITE_DATES_AS_TIMESTAMPS);
+        JsonFactory jfactory = objectMapper.getFactory();
+        JsonGenerator jsonOut = jfactory.createGenerator(System.out, JsonEncoding.UTF8);
+        if (prettyPrint) {
+            jsonOut.useDefaultPrettyPrinter();
+        }
+        return jsonOut;
+    }
+}
